Internalize value object abstraction into its only extension
