+++ src/ApiKey/ApiKey.php
@@ -4,8 +4,25 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\ApiGuard\ApiKey;
 
-use CultuurNet\UDB3\ApiGuard\ValueObjects\StringLiteral;
-
-final class ApiKey extends StringLiteral
+final class ApiKey
 {
+    /**
+     * @var string
+     */
+    protected $value;
+
+    public function __construct(string $value)
+    {
+        $this->value = $value;
+    }
+
+    public function toNative(): string
+    {
+        return $this->value;
+    }
+
+    public function __toString(): string
+    {
+        return $this->toNative();
+    }
 }
